+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### Latest
 -  Adding more general ratio_group argument to puma.Histogram [#67](https://github.com/umami-hep/puma/pull/67)
+- Adding `calc_separation()` to `puma.metrics`, which allows to calculate the separation between two distributions [#27](https://github.com/umami-hep/puma/pull/27)
 - Adding Zenodo link
 
 ### [v0.1.2]